+++ apps/server/src/api/products/(products)/swap/models/token.model.ts
@@ -1,0 +1,4 @@
+import { swapMongoDbConn } from "../../../../../lib/db-connect"
+import { TokenSchema } from "../schemas/token.schema"
+
+export const TokenModel = swapMongoDbConn.model("token", TokenSchema)