+++ apps/server/src/api/products/(products)/swap/models/transaction.model.ts
@@ -1,0 +1,4 @@
+import { swapMongoDbConn } from "../../../../../lib/db-connect"
+import { TransactionSchema } from "../schemas/transaction.schema"
+
+export const TransactionModel = swapMongoDbConn.model("transaction", TransactionSchema)