+++ apps/server/src/api/products/kvstore/models/kv.model.ts
@@ -1,4 +1,0 @@
-import { kvstoreMongoDbConn } from "../../../../lib/db-connect"
-import { KvSchema } from "../schemas/kv.schema"
-
-export const KvModel = kvstoreMongoDbConn.model("kv", KvSchema)