+++ apps/server/src/api/products/nftstudio/models/transaction.model.ts
@@ -1,4 +1,0 @@
-import { nftstudioMongoDbConn } from "../../../../lib/db-connect"
-import { TransactionSchema } from "../schemas/transaction.schema"
-
-export const TransactionModel = nftstudioMongoDbConn.model("transaction", TransactionSchema)